+++ src/App.js
@@ -2,7 +2,6 @@ import React from 'react';
 import Todo from '../src/components/Todo.js'
 import './App.css'
 
-//!14- fontawesome ı ekleyelim
 
 function App() {
   return (
