+++ src/components/Todo.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
-import ListTodos from './ListTodos'
+import TodoList from './TodoList'
 
  class Todo extends React.Component {
-     //! 2-constructor oluşturuldu altında props ve state alıdı
+     //! 2-Constructor was created and received props and state.
     constructor(props) {
         super(props);
         
@@ -14,12 +14,12 @@ import ListTodos from './ListTodos'
             }
         }
 
-        //! 6- bind edelim
+        //! 6- Bind operation done.
         this.handleInput = this.handleInput.bind(this)
         this.addTodo = this.addTodo.bind(this)
     }
 
-    //! 4-metot tanımla
+    //! 4-The function (for input) is coded.
     handleInput(e) {
         this.setState ({
             currentItem: {
@@ -29,13 +29,13 @@ import ListTodos from './ListTodos'
         })
     }
 
-    //! 5- add tanımla
+    //! 5- The function (for form) is coded.
     addTodo(e) {
         e.preventDefault();
 
         const newTodo = this.state.currentItem
         console.log(newTodo);
-        //! 7-Eklemeye ilk adım ardından listTodos ları yarat
+        //! 7-Item 5 has been developed.
         if(newTodo.text !== '') {
             const NewItems = [...this.state.items, newTodo];
             this.setState({
@@ -50,17 +50,17 @@ import ListTodos from './ListTodos'
 
     render() {
         return (
-            //! 1-yapı oluşturuldu
+            //! 1-The main structure has been created. Worked css for app view. (return in render func inside Todo.js)
             <div className="todo-container">
                 <div className="todo-body">
                 <h1 className="todo-h1">What's Your Plan?</h1>
                 <form id="todo-form" onSubmit={this.addTodo}>
-                      {/*! 3-input a value ve button a func eklendi */}
+                      {/* //! 3-Added input a value and onChange function. Added onSubmit to the form. And the component called in the app  */}
                     <input type="text" placeholder="Write a task!" value={this.state.currentItem.text} onChange={this.handleInput} />
                     <button type="submit">Add Todo</button>
                 </form>
                 {/* 12-Comp Added */}
-                <ListTodos items={this.state.items}></ListTodos>
+                <TodoList items={this.state.items}></TodoList>
                 </div>
             </div>
         )
