+++ src/components/TodoList.css
@@ -1,4 +1,4 @@
-/* 13 - add sayfa dizaynı */
+/* 13 Add css coded for list area design. */
 
 .list {
     color: #fff;
